+++ src/git-source-settings.ts
@@ -50,6 +50,16 @@ export interface IGitSourceSettings {
   fetchDepth: number
 
   /**
+   * Fetch tags, even if fetchDepth > 0 (default: false)
+   */
+  fetchTags: boolean
+
+  /**
+   * Indicates whether to use the --progress option when fetching
+   */
+  showProgress: boolean
+
+  /**
    * Indicates whether to fetch LFS objects
    */
   lfs: boolean